Efficiency Update
Make Reef Exploration slightly faster to load.

The sand colours are made global, rather than making a new one each time.

Sand above global y = 0 is statically one colour, rather than having it lerp through the same colour.

A global down vector3 is defined so that a new vector3 is not created every time.

The raycast to define normals is limited to 5 units and now exclusively intersects through the terrain, rather than everything in the scene.

Moveable objects no longer need a raycast to define a normal.

The smoothing function is now a sine ease in out function, rather than cubic ease in out, as was mistakenly stated in the video.

The normal rotations of the corals are now correct... i hope...